requirements.txt
decorator==4.3.0
lxml==4.2.5
networkx==2.2
python-docx==0.8.7
python-xlib==0.23
PyYAML==3.13
ruamel.yaml==0.15.81
six==1.12.0
